Testing: Disable 2018 answer flags
The TC tests are still using 2018 answer reproducibility flags, which
fail when aggressive initialization of allocatable arrays is enabled.
In at least one case, the `mask_z` variable of the ALE sponge is
incompletely initialized when `DEFAULT_2018_ANSWERS` is set.  When
unset, the array is fully initialized.

This patch disables the 2018 answer flags and restores reproducibility
on such platforms.