(*)Offset G%bathyT by -G%Z_ref
  Added code to adjust the value of G%bathyT in the ocean_grid_type by -G%Z_ref,
with changes scattered throughout the code to compensate.  This does not impact
the value of bathyT in the dyn_hor_grid.  Using a non-zero value of
REFERENCE_HEIGHT leads to mathematically equivalent solutions that differ at
round-off but are demonstrably similar for values of REFERENCE_HEIGHT ranging
from -1e4 m to 1e4 m, and in cases that do not use the less exact
..._2018_ANSWERS algorithms the solutions are qualitatively equivalent for
values ranging from -1e8 m to 1e8 m.  (For these more extreme values, the 64-bit
roundoff in the free surface height calculation starts to become physically
significant at of order 0.1 mm.)  This new capability is useful for developing
wetting and drying and for identifying algorithmic shortcomings, but because
answers are not bitwise identical for various reference heights, it is not such
a good candidate for fully automated testing.  By default, all answers are
bitwise identical and all output files are the same.