ALE sponge: Only update fields on uv over masks
Currently, fields from `horiz_interp_and_extrap_tracer` are interpolated
onto [uv] points under the assumption that halos have sensible values.
This is generally true due to the preceding `pass_var` call, but it may
not be valid if the halo is not updated, e.g. along a boundary or
land-masked tile.

The mask is designed to be set to zero when this happens, but it could
still result in an assignment of an uninitialized value to the fields
(`sp_val_[uv])`.  Although the value is not used, it can produce
compiler warnings and errors under strict debugging builds.

This patch moves the calculation of `sp_val_[uv]` so that it is only
conditionally computed when the mask is set to 1.