Infra: MOM_read_data rename to read_field
This patch removes the `read_data_infra` alias in `MOM_io` to the
infra's identically named `MOM_read_data`, and instead renames the infra
function to `read_field`.

This is primarily done to avert the inability of the PGI compiler to
resolve the alias in the namespace.

It also provides a slightly more consistent namespace:

1. It resembles the `write_field` functions in the infra layer.

2. It avoids reuse of the `read_data` name, used in FMS.

3. It allows the MOM framework layer to preserve its `MOM_` suffix,
   as expected for a layer intended for use within MOM.

4. `read_field` sheds any explicit reference to `MOM_`, helping to
   identify the infra layer as a generic interface to its framework.