+(*)Add ALTERNATE_FIRST_DIRECTION
  Added a new runtime parameter, ALTERNATE_FIRST_DIRECTION, to cause the first
direction in the directionally split operators (such as the continuity solver,
the barotropic solver, and the tracer advection) to alternate with every dynamic
timestep.  This includes changes to record this direction in the restart file,
so that the model will reproduce across restarts.  In addition a previously
uninitialized logical in offline_advection_layer is now being initialized
similarly to other similar variables in the MOM_offline_main module.  By
default, all answers in the MOM6-examples are bitwise identical, but there is a
new runtime parameter and a new non-mandatory field in the MOM6 restart files,
and there might be changes in offline tracer advection runs.