Testing: Configurable list of test targets
`make all` (the default rule) builds the executables specified by
`BUILDS`.  This was hard-coded, but can be promoted to a user-defined
configuration for user-defined builds.

This should be seen as a simple alias to `make build/${b}/MOM6`.

The `repro` build was also incorrectly in the list, even though it was
also conditionally added.  It has been removed from the default list.

A similar modification was made to CONFIGS, which select the "tc"
experiments.  The default is still to do a `tc*` glob.

Documentation was also added.