+Add BAROTROPIC_TIDAL_SAL_BUG to fix a tide bug
  Added a runtime flag, BAROTROPIC_TIDAL_SAL_BUG, to fix a sign error in the
tidal self-attraction and loading anomalies in the barotropic solver when tides
are enabled.  The default is to keep the previous bug so that answers do not
change, but this default will be changed after solutions have been corrected.
This commit partly addresses MOM6 issue #1496, but it should only be considered
to be properly handled once the default has been changed to avoid this bug.
This commit will change the MOM_parameter_doc files in cases where this bug
matters, but by default all answers are bitwise identical.