Entrain diffusive pointer removal
This removes just about all of the entrain_diffusive pointers (except
the diag_ctrl).

There is a very minor incongruity with `just_read_params`, which was
originally used to deallocate the CS, which might alter some removed
`if(associated(CS))` checks.  But it seems this is not really a problem,
since the calls to entrainment_diffusion() are inside regions
unreachable when this flag (from `CS%useALEalgorithm`) is true.