+Add new parameter FATAL_INCONSISTENT_RESTART_TIME
  Add the new runtime parameter FATAL_INCONSISTENT_RESTART_TIME, which if true
causes the model to compare the restart time read from a restart file with any
value provided as a time_in to MOM_initialize_state and issue a fatal error if
they differ.  In ocean-only mode, this input time is read from a ocean_solo.res
file, following FMS behavior.  The default value simply uses the specified time,
replicating the previous behavior.  If set to true, this would prevent a problem
with the time-history that recently occurred in a series of high-resolution runs
that were shared between several groups, where the nominal times were repeated.
All answers are bitwise identical, but many MOM_parameter_doc files have a new
entry.