+Reverted default for PARALLEL_RESTARTFILES
  Reverted the default for PARALLEL_RESTARTFILES to False because the TC restart
testing was having problems with this set to True.  This is surprising because
PARALLEL_RESTARTFILES = True has been used in production runs for many years now
without any indication that MOM6 fails to reproduce across restarts in this
mode, so this could be an issue with the TC testing.  In the mean time,
reverting this default will allow the other changes to be accepted while this
curious behavior is explored separately.  Several comments related to
PARALLEL_RESTARTFILES were also updated for consistency. All answers are bitwise
identical.