Refactored solo_driver/MOM_driver.F90
  Refactored solo_driver/MOM_driver.F90 to move logically self-contained blocks
of code into separate subroutines within this file to improve the readability of
the code and to reduce the number of global variables.  This started out as an
exercise to explore the use of the F2008 block / end block construct to reduce
the scope of variables, but the code ended up being cleaner just using
traditional subroutines contained in this file.  All answers are bitwise
identical and all output files are unaltered.