FMS1: Don't create IO domains for single-PE domain
FMS mpp domain creation is done in the `clone_MD_to_d2D` function, and
currently an IO domain is always created within the MPP domain.

This has caused problems with single-PE runs in FMS1, where the
`write_field` logic was not able to reach the part which removes halo
data if an IO domain was present, and halo data was being written to the
restart files.

This issue arose when `PARALLEL_RESTARTFILES` was set to `True` for
single-PE tests.

This does not appear to be a problem with FMS2, and no action is needed
by the FMS team.