Always create single-threaded files in 1 PE runs
  Modified create_file to always specify that the threading will be for a single
PE to do the writing to a single file if there is only a single ocean PE, to
avoid some incorrect behavior inside of the FMS IO modules.  This can fix
restart problems in some single-processor cases with an inconsistent setting of
the optional threading argument, but in all cases that ran correctly before, all
answers are bitwise identical.