Use FMS2 `file_exists`, remove domain args
This patch uses the FMS2 `file_exists` function when using the FMS2
infra.  Previously, the FMS1 version of this function was being used.

This patch also removes the `mpp_domain` and `no_domain` arguments from
the direct `file_exist` calls, which were not used by any known MOM6
configurations.  (Nor would we expect them to be, since explicit
references to FMS should not exist outside of the infra layer.)

Since FMS2 does not use these arguments, their removal also creates a
more meaningful interface between the two frameworks.

Motivation:

An issue with the FMS1 `file_exists` under the FMS2 infra was discovered
in the UFS model on Hera.  It was only reproducible in submitted jobs,
and not for interactive jobs, and only with the GCC 9.2 compiler.
(Other GCC versions were not tested.)

One potential explanation is that it is related to the `save` attribute
of the domain pointer, `d_ptr`.  In the case above, `d_ptr` pointed to
the MOM input domain for the failed cases.  For the other working
cases, `d_ptr` pointed to a `NULL()` value and behavior was normal.

It is possible that `d_ptr` is inconsisently updated when FMS1 and FMS2
IO operations are used together, which should probably be considered
undefined behavior.