cpu_clock_id: synchro_flag arg changed to logical
This patch modifies the `cpu_clock_id` interface so that the
`synchro_flag` argument is converted into a platform-agnostic logical
flag.

The current implementation requires the synchronization flag to be
defined using FMS norms (zero-bit) and also would force users to follow
FMS predefined flags for other values.

This patch changes the sync flag to a logical, and modifies the default
flag to enable synced clocks (based on `clock_flag_default`).

`synchro_flag` is also renamed to `sync` for simplicity.