Eliminate GET_ALL_PARAMS in hor_visc_init (#1536)
* Eliminate GET_ALL_PARAMS in hor_visc_init

  Added do_not_log arguments to get_param calls in MOM_hor_visc.F90 that are
only used conditionally, and eliminated the unlogged GET_ALL_PARAMS runtime
parameter and get_all variable in hor_visc_init().  By design, all logging of
parameters after this commit is identical to before, even for variables that are
inactive and therefore should not be logged.  In several places, there were some
problems, mostly with the GME code, that have been noted in comments marked with
'###'.  Also cleaned up the code alignment and eliminated unneeded temporary
variables in a few places in hor_visc().  All solutions are bitwise identical,
and no output is changed.

* Restore temporary variables

  Undid changes that eliminated temporary variables to facilitate performance
profiling, and restored the "Knuth" convention about the placement of the line
breaks relative to "+" in these expressions.  Nothing changes.