+Add G%Rad_Earth_L and related code simplification
  Add a dimensionally rescaled variable with the radius of the Earth in [L ~> m]
to the ocean_grid_type and dyn_hor_grid type to expand dimensional consistency
testing.  For now it exists along-side the previous variable with the radius of
the Earth in [m], which might be used in other drivers or user code that we are
not aware of.  All instances in the MOM6 code that were using the unscaled
variable G%Rad_Earth were changed to use the dimensionally scaled variable
G%Rad_Earth_L.  In addition, the code for setting up the grid metrics was
simplfied by completing the dimensional rescaling of its internal variables
and some grid rescaling factors that were no longer in use were eliminated.
As a part of this change, a unit_scale_type argument was added to the interface
for USER_initialize_tracer.

  In addition, a number of comments were changed to clearly distinguish between
dimensionally scaled variables that scale back to a nondimensional variable that
can take on a range of values (using the syntax [... ~> nondim]), as opposed to
variables that are only a combination of dimensional scaling factors that
resolve back to 1 in MKS units (using the syntax [... ~> 1]).

  All answers are bitwise identical, but there is one minor interface change
and there are new elements in several widely used transparent types.