(*)Fix rescaling when SPLIT_BOTTOM_STRESS = True
  Corrected the dimensional rescaling of the explicit bottom stresses that are
passed to btstep() when SPLIT_BOTTOM_STRESS = True.  Without this correction,
solutions would not pass the dimensional consistency tests in this case.
Thankfully, the parameter setting in question does not seem to be widely used,
and even if it is, answers will not change when Z_RESCALE_POWER = 0.  All
answers and output in the MOM6-examples test suite are bitwise identical.