(*)Refactored and fixed pseudo_salt_tracer code
  Refactored the pseudo_salt_tracer code to avoid using fluxes%netSalt, instead
calculating the surface fluxes directly from fluxes%saltFlux, in preparation to
get rid of the largely redundant fluxes%netSalt element, but also to correct for
the fact that fluxes%netSalt does not have proper scaling to pass dimensional
consistency testing.  Also made the CS%diff element of pseudo_salt_tracer_CS
into an allocatable instead of it being a pointer, and changed the tests for
whether it is allocated appropriately.  Several unused variables were
eliminated, and some comments corrected or reformated.  This does not impact
solutions, but the pseudo_salt and pseudo_salt_diff diagnostics are now
invariant to dimensional rescaling, whereas before they were not.  All solutions
and output in MOM6-examples are bitwise identical.