Expanded comments describing src/core code (#13)
Added comments describing all of the real variables in the src/core directory
that did not previously have them, and corrected multiple spelling error in this
same directory.  This PR also updates the description of the MOM6 code structure
in MOM.F90 and adds dimensional descriptions to the commends describing a number
of variables.  Descriptions of the units of some variables were also added or
corrected in a number of the src/tracer files.  Changes are largely restricted
to comments describing varibles, although a few unused variables were removed.
One comment added to MOM_dynamics_split_RK2.F90 notes a possible memory leak,
due to a missing deallocate call, but the actual correction of this leak will be
delayed to a later, much more targeted commit.  All answers and output are
bitwise identical.