Testing: Codecov uploader update
This patches updates the codecov uploader to the new version, replacing
the bash uploader to be phased out in 2022.

The uploader URL is now a configuable variable, and the coverage scripts
have been modified to accommodate changes in the uploader.

Under this new uploader, the coverage reports must be generated locally,
rather than relying on the bash script to call gcov.  So we now generate
these scripts based on the *.gcda output.