Bugfix: cpu clock sync error
This patch fixes an error in the `sync` flag of the cpu clocks.

Previously, we would set the sync bit of a flag based on the presence of
sync, rather than testing if the value was true.  This would cause
potential hangs in any clock that set `sync`, including `.false.`.

This patch correctly replaces the single `ibset` call with an if-block
to either `ibset` or `ibclr`.