+Remove clocks inside of j-loops
  Removed clocks that were being called from inside of j-loops in two modules.
These are inefficient and can cause the model to hang in some cases if used, and
there are better ways to get timing information at this level.  If there is
interest in the timing breakdown at this level, the code should be restructured
to move the key blocks outside of the j-loops.  The run-time parameter
ALLOW_CLOCKS_IN_OMP_LOOPS is no longer being used so it is now obsoleted. All
answers are bitwise identical, but there are changes to some MOM_parameter_doc
files.