+Set find_salt_root if SHELF_THREE_EQN = .False.
  Set find_salt_root even if SHELF_THREE_EQN = .False. to avoid using an
uninitialized logical to determine which parameters are logged.  Without this
the contents of some MOM_parameter_doc.all files could depend on the state of
uninitialized memory and was compiler dependent in some cases.  All answers are
bitwise identical, but in some cases the contents of MOM_parameter_doc files
could be corrected.