(*)Provide US arguments to 4 existing calls
  Provide optional US arguments to 4 existing calls to set_grid_metrics() and
MOM_initialize_topography(), enabling these arguments to become mandatory in the
next commit.  In some cases this requires a call to unit_no_scaling_init() or
the removal of a call to rescale_dyn_horgrid_bathymetry().  In addition, a
mis-scaled value was corrected in the initialization of the ODA control
structures private thickness array; this latter issue is not a problem for
Boussinesq models without dimensional rescaling (i.e. the tested cases), but
could change answers in other cases with data assimilation.  All answers in
the MOM6-examples or TC test cases are bitwise identical.