Warn if opacity_from_chl is called without fluxes
  Issue a warning with a helpful message if opacity_from_chl is called with no
shortwave fluxes, and added logical tests to avoid a segmentation fault later in
this routine.  This should not happen, as it makes no sense, but it was
occurring with the offline tracer code, and can be avoided by setting
PEN_SW_NBANDS=0 if there are no shortwave fluxes to penetrate.  Also turned the
real dimensional parameter op_diag_len into a variable and set it immediately
before where it is used.  Many spelling errors were also corrected in
MOM_opacity.F90.  All answers are identical in cases that ran before.