Clean up non-standard syntax and whitespace
  Eliminated a number of instances of non-standard syntax from that described in
https://github.com/NOAA-GFDL/MOM6/wiki/Code-style-guide.  The changes include
enforcing the MOM6-standard 2-point indentation convention, replacing 'if(A)'
with 'if (A)', and changing logical comparison syntax like '.gt.' to '>' or
'.eq.' to '=='.  An old commented out block of code for debugging (detected by
its use of non-standard syntax) was also eliminated.  All answers and output are
bitwise identical.