(*)Correct memory declarations in MOM_regridding
  Corrected memory declarations in MOM_regridding.F90 in cases where the
vertical size of the input and output grids are not the same.  Although this is
not know to have caused any particular problems, these inconsistencies could
lead to segmentation faults in cases where the target grid (e.g., diagnostic
output) is larger than the input grid (e.g., the model's native grid).  In some
cases, certain grid generation options were only written to work with the same
size of input and output grids, and error handling has been added to these cases
to gracefully bring down the model if they are used with different grid sizes.
All answers are bitwise identical in the MOM6-examples test suite, but it is
conceivable that this could correct subtle (memory-related) issues in some
configurations.