read_variable_2d modified to accept 3 or 4 dims
The read_variable_2d function was previously configured to only run if
the start and nread arrays matched the size of the field they were
accessing.  This was incompatible with the history of the function,
which had previously required a fourth time axis (of one record), then
was later modified to not require this axis.  As a result, there are now
files in use both with and without a time axis.

This patch relaxes this check to ensure that the read is quasi-2d, i.e.
the first two axes can read a segment of a 2d field, but will now
reshape the start and nread arrays to match the field being read.  Some
additional checks are also added to ensure that it only reads one 2d
slice.